11/06/2025 - Home
I got the Saving and Loading function completed, and many other elements are starting to work better together. The current objective to to keep working through my game, and when I run into a feature that isnt fully complete or polished, I will work on finishing that until it is done, then move to the next thing.