16/06/2025 - Home
Finished a variety of systems, and now it is fully working. I want to expand the inventory system a bit and make it more detailed, and make a boss fight. I still need to put the story line in. I also started adding in some of the Charts etc, to put in the markdown file, and downloading them from where they where first made.